applied updates
