+++ composer/datasets/streaming/writer.py
@@ -6,17 +6,19 @@
 
 import gzip as gz
 import os
+import urllib
 from io import BufferedWriter
 from types import TracebackType
 from typing import Dict, Iterable, List, Optional, Tuple, Type, Union
 
 import numpy as np
 from tqdm import tqdm
-from upload import get_object_store
 
 from composer.datasets.streaming.format import (StreamingDatasetIndex, get_compression_scheme_basename,
                                                 get_index_basename, get_shard_basename, sample_dict_to_bytes)
 from composer.utils.object_store.object_store import ObjectStore
+from composer.utils.object_store.s3_object_store import S3ObjectStore
+from composer.utils.object_store.sftp_object_store import SFTPObjectStore
 
 __all__ = ['StreamingDatasetWriter']
 
@@ -84,6 +86,7 @@ class StreamingDatasetWriter(object):
         fields: (List[str]): The fields to save for each sample.
         shard_size_limit (int): Maximum shard size in bytes. Default: ``1 << 24``.
         compression (str, optional): Compression algorithm and optional compression level. Currently supported: 'gz', 'gz:[1-9]' or None. Defaults to ``None``.
+        remote ((ObjectStore | str), optional): Remote location to upload dataset. Can be an ObjectStore, a string, or (default) None.
     """
 
     default_compression = None
@@ -93,8 +96,7 @@ class StreamingDatasetWriter(object):
                  fields: List[str],
                  shard_size_limit: int = 1 << 24,
                  compression: Optional[str] = default_compression,
-                 remote: Union[ObjectStore, str, None] = None,
-                 timeout: Optional[float] = None) -> None:
+                 remote: Optional[Union[ObjectStore, str]] = None) -> None:
         if len(fields) != len(set(fields)):
             raise ValueError(f'fields={fields} must be unique.')
         if shard_size_limit <= 0:
@@ -117,7 +119,7 @@ class StreamingDatasetWriter(object):
         # compression scheme for shards
         self.compression_scheme, self.compression_level = _parse_compression_args(compression)
 
-        self.remote = _parse_remote(remote, timeout)
+        self.remote = _parse_remote(remote)
 
     def _create_binary_file(self, fname: str) -> Union[BufferedWriter, gz.GzipFile]:
         """opens a (potentially compressed) file in binary mode"""
@@ -219,12 +221,68 @@ class StreamingDatasetWriter(object):
         self.finish()
 
 
-def _parse_remote(remote: Union[ObjectStore, str, None], timeout: Optional[float]) -> Optional[ObjectStore]:
+def _parse_remote(remote: Union[ObjectStore, str, None]) -> Optional[ObjectStore]:
     if isinstance(remote, str):
-        return get_object_store(remote, timeout)
+        return _get_object_store(remote)
     elif isinstance(remote, ObjectStore):
         return remote
     elif remote is None:
         return None
     else:
         raise ValueError('Bad argument for remote')
+
+
+def _get_object_store(remote: str) -> ObjectStore:
+    """Use the correct download handler to download the file
+
+    Args:
+        remote (Optional[str]): Remote path (local filesystem).
+        timeout (float): How long to wait for file to download before raising an exception.
+    """
+    if remote.startswith('s3://'):
+        return _get_s3_object_store(remote)
+    elif remote.startswith('sftp://'):
+        return _get_sftp_object_store(remote)
+    else:
+        raise ValueError('unsupported upload scheme')
+
+
+def _get_s3_object_store(remote: str) -> S3ObjectStore:
+    obj = urllib.parse.urlparse(remote)
+    if obj.scheme != 's3':
+        raise ValueError(f"Expected obj.scheme to be 's3', got {obj.scheme} for remote={remote}")
+    bucket = obj.netloc
+    object_store = S3ObjectStore(bucket=bucket)
+    return object_store
+
+
+def _get_sftp_object_store(remote: str) -> SFTPObjectStore:
+    url = urllib.parse.urlsplit(remote)
+    # Parse URL
+    if url.scheme.lower() != 'sftp':
+        raise ValueError('If specifying a URI, only the sftp scheme is supported.')
+    if not url.hostname:
+        raise ValueError('If specifying a URI, the URI must include the hostname.')
+    if url.query or url.fragment:
+        raise ValueError('Query and fragment parameters are not supported as part of a URI.')
+    hostname = url.hostname
+    port = url.port
+    username = url.username
+    password = url.password
+
+    # Get SSH key file if specified
+    key_filename = os.environ.get('COMPOSER_SFTP_KEY_FILE', None)
+    known_hosts_filename = os.environ.get('COMPOSER_SFTP_KNOWN_HOSTS_FILE', None)
+
+    # Default port
+    port = port if port else 22
+
+    object_store = SFTPObjectStore(
+        host=hostname,
+        port=port,
+        username=username,
+        password=password,
+        known_hosts_filename=known_hosts_filename,
+        key_filename=key_filename,
+    )
+    return object_store
